Still not working

+++ myMenus/inventory.hpp
@@ -11,7 +11,6 @@ using namespace std;
 
 class Inventory{
 
-	int choice;
 public:   
 	Inventory() {}
  
@@ -23,7 +22,7 @@ public:
 	}   
 
 	virtual void menuFunc(Character* c){
-			
+			int choice;			
 		do{
 			printMenu(c);
 			cout << "Select an item or exit inventory: ";
@@ -33,30 +32,40 @@ public:
 				break;
 			}
 
-			if(c->isHolding() == true && c->getHand()->getName() == c->getItem(choice)->getName()){
-				std::string choice2;
-				cout << c->getHand()->getName() << " is equipped. Do you want to unequip?(Y/N): ";
-				cin >> choice2;
-				cout << endl;
-				if(choice2 == "Y" || choice2 == "y"){
-					c->unequipItem(choice);
-				} 
-			} else if(c->isHolding() == false){
-				std::string choice3;
-				cout << "You have selected " << c->getItem(choice)->getName() << ". EQUIP or DELETE?(E/D): ";
-				cin >> choice3;
-				cout << endl;
-				if(choice3 == "D" || choice3 == "d"){
-					c->deleteItem(choice);
-					cout << "Reloading items..." << endl;
+			if(choice > 3){
+				cout << "Invalid selection. Try again." << endl;
+			} 
+
+			if(choice <= 3 && choice > 0){
+				if(c->isHolding() == true && c->getHand()->getName() == c->getItem(choice)->getName()){
+					std::string choice2;
+					cout << c->getHand()->getName() << " is equipped. Do you want to unequip?(Y/N): ";
+					cin >> choice2; 
+
+					if(choice2 == "Y" || choice2 == "y"){
+						c->unequipItem(choice);
+					} else {
+
+					} 
+				} else if (c->isHolding() == false){
+					std::string choice3;
+					cout << "You have selected " << c->getItem(choice)->getName() << ". EQUIP or DELETE?(E/D): ";
+					cin >> choice3;
 					cout << endl;
-				} else if (choice3 == "E" || choice3 == "e"){
-					c->equipItem(choice);
+					if(choice3 == "D" || choice3 == "d"){
+						c->deleteItem(choice);
+						cout << "Reloading items..." << endl;
+						cout << endl;
+					} else if (choice3 == "E" || choice3 == "e"){
+						c->equipItem(choice);
+					} else {
+
+					}
+
 				}
+			}	
 
-			}
-			
-
+		
 		}while (choice != 0);
 	}
 };
